requirements.txt
Django==3.2.17
psycopg2-binary==2.9.5
djangorestframework==3.14.0
django-filter==22.1